adição dos atributos names e email no jsonWebToken

+++ src/modules/donations/useCases/deleteDonation/DeleteDonationUseCase.ts
@@ -1,13 +1,24 @@
 import { validationResponse } from "../../../../types";
 import { Donations } from "../../entities/Donations";
 import { IDonationsRepository } from "../../repositories/IDonationsRepository";
+import { checkBody } from "./DeleteDonationCheck";
 import { DeleteDonationProps } from "./DeleteDonationController";
 
 class DeleteDonationUseCase {
     constructor(
         private donationsRepository: IDonationsRepository) {}
 
-    async execute(donationID:Donations["id"], donationData: DeleteDonationProps): Promise<Donations | validationResponse> {
+    async execute(donationID:Donations["id"], donationData: DeleteDonationProps): Promise<validationResponse> {
+
+        const bodyValidation = await checkBody(donationData)
+
+        if (bodyValidation.isValid === false) {
+            return ({
+                isValid: false,
+                statusCode: 403,
+                errorMessage: bodyValidation.errorMessage,
+            })
+        }
         
         const deletedDonation = await this.donationsRepository.deleteDonation(donationID, donationData)
